MVC PATERN
